Implement the VideoCaptureDevicePxcWin based on PXCCapture API
PXCCapture is the Intel PCSDK API to access video streams including color
and depth streams on Windows OS.

VideoCaptureDevicePxcWin is the capture foundation to bring PerC capability
to Crosswalk. It serves:
- Capture color images for standard getUserMedia.
- Capture depth images for experimental 'depth' extension of getUserMedia.
- Avoid camera accessing confliction with other PXCCapture API clients.

To build it in, please install the Intel PCSDK and append gyp definition
"-Duse_pxc_capture=1" when generating the project.